replace references using providesModule
